naming enhancements and remove gmailauth
